# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "Documentación completa de la API"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/usuarios": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Obtener todos los usuarios de una agencia",
        "description": "Permite obtener la lista completa de usuarios (tanto admins como usuarios regulares) de una agencia",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-agencia-id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID de la agencia para filtrar usuarios"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "resultado": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "nombre": {
                            "type": "string"
                          },
                          "apellido": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string",
                            "format": "email"
                          },
                          "rol_id": {
                            "type": "integer"
                          },
                          "agencia_id": {
                            "type": "integer"
                          },
                          "activo": {
                            "type": "boolean"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No se encontraron usuarios para la agencia especificada"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
...
    "/api/usuarios/superadmin/spy_mode": {
      "put": {
        "tags": [
          "Usuarios SuperAdmin Spy Mode"
        ],
        "summary": "Asigna un ID de agencia al superadmin (Spy Mode)",
        "description": "Permite al superadmin acceder temporalmente como un usuario de agencia (agencyuser/agencyadmin) asignándole un ID de agencia.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "agencia_id",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": true
            },
            "description": "ID de la agencia a asignar al superadmin. Si es null, se elimina la asignación."
          }
        ],
        "responses": {
          "200": {
            "description": "ID de agencia asignado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Cliente modificado exitosamente"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "uuid-usuario"
                        },
                        "agencia_id": {
                          "type": "integer",
                          "nullable": true,
                          "example": 123
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID de agencia no válido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "ID de agencia no válido"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "ID de agencia no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "ID de agencia no encontrado"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Error interno del servidor"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}